Jars declared in policy file. Related to #387
