+++ aggregator/pom.xml
@@ -79,7 +79,7 @@
                 <artifactId>maven-shade-plugin</artifactId>
                 <version>2.3</version>
                 <configuration>
-                    <finalName>${project.name}-${project.version}-plugin</finalName>
+                    <finalName>${project.name}-plugin</finalName>
                     <filters>
                         <filter>
                             <artifact>*:*</artifact>
