--- test-at/src/test/resources/policies/ISocket-OParquet.json
@@ -1,107 +1,0 @@
-{
-  "name": "ATParquetOutput",
-  "sparkStreamingWindow": 2000,
-  "rawData": {
-    "enabled": "false",
-    "partitionFormat": "day",
-    "path": "myTestParquetPath"
-  },
-  "checkpointing": {
-    "path": "checkpoint",
-    "timeDimension": "minute",
-    "granularity": "minute",
-    "interval": 30000,
-    "timeAvailability": 60000
-  },
-  "inputs": [
-    {
-      "name": "in-socket",
-      "type": "Socket",
-      "configuration": {
-        "hostname": "localhost",
-        "port": "10666"
-      }
-    }
-  ],
-  "transformations": [
-    {
-      "name": "morphline-parser",
-      "order": 0,
-      "type": "Morphlines",
-      "outputFields": [
-        "product",
-        "price"
-      ],
-      "configuration": {
-        "morphline": {
-          "id": "morphline1",
-          "importCommands": [
-            "org.kitesdk.**"
-          ],
-          "commands": [
-            {
-              "readCSV": {
-                "separator": ",",
-                "columns": [
-                  "product",
-                  "price"
-                ],
-                "trim": "true",
-                "charset": "UTF-8",
-                "quoteChar": "\"",
-                "commentPrefix": "#"
-              }
-            }
-          ]
-        }
-      }
-    },
-    {
-      "name": "timestamp",
-      "order": 1,
-      "type": "DateTime",
-      "outputFields": [
-        "timestamp"
-      ],
-      "configuration": {
-        "inputFormat": "autoGenerated"
-      }
-    }
-  ],
-  "cubes": [
-    {
-      "name": "testCube",
-      "dimensions": [
-        {
-          "name": "product",
-          "field": "product"
-        }
-      ],
-      "operators": [
-        {
-          "name": "sumoperator",
-          "type": "Sum",
-          "configuration": {
-            "inputField": "price"
-          }
-        },
-        {
-          "name": "avgoperator",
-          "type": "Avg",
-          "configuration": {
-            "inputField": "price"
-          }
-        }
-      ]
-    }
-  ],
-  "outputs": [
-    {
-      "name": "out-parquet",
-      "type": "Parquet",
-      "configuration": {
-        "path": "/tmp/sparkta/outputs/parquet"
-      }
-    }
-  ]
-}
